Ensure all import-errors caused by disabled cli extra are caught
A recent commit accidentally imported a rich component at the top-level,
outside of the import-guard. This was bound to happen with IDEs adding
new imports at the top (understandably so).

To ensure this does not bite us again, this commit slightly restructures
the CLI package to encapsulate the CLI code in a separate file that is
imported under a guard once.
